type change from long to int for IsKnownMinionID APIs

+++ GW2EIEvtcParser/EIData/ProfHelpers/Engineer/EngineerHelper.cs
@@ -196,8 +196,8 @@ namespace GW2EIEvtcParser.EIData
             }
         }
 
-        private static HashSet<long> Minions = new HashSet<long>();
-        internal static bool IsKnownMinionID(long id)
+        private static HashSet<int> Minions = new HashSet<int>();
+        internal static bool IsKnownMinionID(int id)
         {
             return Minions.Contains(id);
         }
