+++ GW2EIEvtcParser/EIData/ProfHelpers/Mesmer/MirageHelper.cs
@@ -30,8 +30,8 @@ namespace GW2EIEvtcParser.EIData
             new Buff("False Oasis", FalseOasis, Source.Mirage, BuffClassification.Other, BuffImages.FalseOasis),
         };
 
-        private static HashSet<long> Minions = new HashSet<long>();
-        internal static bool IsKnownMinionID(long id)
+        private static HashSet<int> Minions = new HashSet<int>();
+        internal static bool IsKnownMinionID(int id)
         {
             return Minions.Contains(id);
         }
